+++ db/migrate/20260620220609_add_more_ingredients_to_recipes.rb
@@ -1,8 +1,8 @@
 class AddMoreIngredientsToRecipes < ActiveRecord::Migration[8.1]
   def change
-    add_column :recipes, :brown_sugar, :integer
-    add_column :recipes, :baking_soda, :integer
-    add_column :recipes, :chocolate_chips, :integer
-    add_column :recipes, :vanilla, :integer
+    add_column :recipes, :brown_sugar, :float
+    add_column :recipes, :baking_soda, :float
+    add_column :recipes, :chocolate_chips, :float
+    add_column :recipes, :vanilla, :float
   end
 end
